+++ CHANGELOG.rst
@@ -71,6 +71,8 @@ Fixed
 
 - ``vshard-storage`` ``apply_config`` won't change order in ``box.cfg.replication`` (`#1950 <https://github.com/tarantool/cartridge/issues/1950>`_).
 
+- Allow to use ``box.NULL`` as label value.
+
 -------------------------------------------------------------------------------
 [2.7.6] - 2022-08-22
 -------------------------------------------------------------------------------